DataCollectorMicroservice/requirements.txt
flask
grpcio
grpcio-tools
mysql-connector-python
requests
protobuf